refactor(atoms): group form and graphics atoms into subfolders
Organize the flat atoms/ directory by moving two cohesive groups into
dedicated subfolders, each with its own index.ts barrel (mirroring the
existing svgs/ pattern) re-exported from the parent atoms barrel:

- atoms/form/     — Checkbox, Input, Radio, Select, Textarea
- atoms/graphics/ — Logo, LanguageFlag, SocialIcon, svgs/

The top-level `components/atoms` barrel keeps the same public API, so
`import { Input, Logo } from '../../atoms'` consumers are unchanged.
Relative imports were repointed where the move changed depth:
- form/Input -> ../../Icon/Icon; graphics/Logo -> ../../Link
- moved stories -> ../../../ladle
- sibling atoms (Spinner, SplashLoader, LeafDivider) -> ../graphics/svgs
- organisms/molecules deep imports -> ../../atoms/graphics/svgs (etc.)

Ladle nav now groups these under "Atoms / Form" and "Atoms / Graphics".
Stale doc path references updated. tsc / lint / tests all green.
