+++ components/atoms/form/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Form Inputs - Atomic form-control elements
+ *
+ * Barrel export for the form-input atoms grouped under `atoms/form/`.
+ * Re-exported by the parent `components/atoms` barrel.
+ */
+
+export { Input } from './Input'
+export { Textarea } from './Textarea'
+export { Checkbox } from './Checkbox'
+export { Radio } from './Radio'
+export { Select } from './Select'
+export type { InputProps } from './Input'
+export type { TextareaProps } from './Textarea'
+export type { CheckboxProps } from './Checkbox'
+export type { RadioProps } from './Radio'
+export type { SelectProps } from './Select'
