+++ components/atoms/graphics/Logo/index.ts
