+++ components/atoms/graphics/SocialIcon/index.ts
