+++ components/atoms/graphics/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Graphics - Brand & visual atoms
+ *
+ * Barrel export for the brand/identity and decorative graphic atoms grouped
+ * under `atoms/graphics/` (Logo, LanguageFlag, SocialIcon, and the svgs).
+ * Re-exported by the parent `components/atoms` barrel.
+ */
+
+export { Logo } from './Logo'
+export type { LogoProps } from './Logo'
+
+export { LanguageFlag } from './LanguageFlag'
+export type { LanguageFlagProps } from './LanguageFlag'
+
+export { SocialIcon } from './SocialIcon'
+export type { SocialIconProps } from './SocialIcon'
+
+export * from './svgs'
