+++ components/atoms/graphics/svgs/AnimatedLogoSvg.tsx
