+++ components/atoms/graphics/svgs/HeaderIllustrationSvg.tsx
