+++ components/atoms/graphics/svgs/LeafSvg.tsx
