+++ components/atoms/graphics/svgs/LogoSvg.tsx
