+++ components/atoms/graphics/svgs/LotusDotsSvg.tsx
