+++ components/atoms/graphics/svgs/SimpleLeafSvg.tsx
