+++ components/atoms/graphics/svgs/TriangleDecorationSvg.tsx
