+++ components/atoms/graphics/svgs/index.ts
