+++ components/organisms/FormBuilder/FormBuilder.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react'
 import { useForm, UseFormRegister } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import type { ZodObject, ZodRawShape } from 'zod'
-import { Input } from '../../atoms/Input'
-import { Textarea } from '../../atoms/Textarea'
-import { Select } from '../../atoms/Select'
-import { Checkbox } from '../../atoms/Checkbox'
+import { Input } from '../../atoms/form/Input'
+import { Textarea } from '../../atoms/form/Textarea'
+import { Select } from '../../atoms/form/Select'
+import { Checkbox } from '../../atoms/form/Checkbox'
 import { Button } from '../../atoms/Button'
 import { FormField } from '../../molecules/FormField'
